requirements.txt
requests
pywhatkit
python-dotenv
pyautogui